DoceCostura/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==4.2.4
django-cors-headers==4.3.0
django-filter==25.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
idna==3.10
pillow==11.2.1
PyJWT==2.10.1
python-dotenv==1.0.0
pytz==2025.2
requests==2.31.0
sqlparse==0.5.3
typing_extensions==4.13.0
urllib3==2.3.0
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==4.2.4
django-cors-headers==4.3.0
django-filter==25.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
idna==3.10
mercadopago==2.3.0
pillow==11.2.1
PyJWT==2.10.1
python-dotenv==1.0.0
pytz==2025.2
requests==2.31.0
sqlparse==0.5.3
stripe==12.0.1
typing_extensions==4.13.0
urllib3==2.3.0
